ast building.
yay! first test passes!

+++ branches/branch-dev-new-factory-2/test/nodes/PKCompositeNode.m
@@ -10,7 +10,7 @@
 
 @implementation PKCompositeNode
 
-- (int)type {
+- (NSUInteger)type {
     return PKNodeTypeComposite;
 }
