+++ branches/branch-dev-new-factory-2/test/nodes/PKConstantNode.m
@@ -10,7 +10,7 @@
 
 @implementation PKConstantNode
 
-- (int)type {
+- (NSUInteger)type {
     return PKNodeTypeConstant;
 }
