+++ branches/branch-dev-new-factory-2/test/nodes/PKDelimitedNode.m
@@ -44,7 +44,7 @@
 }
 
 
-- (int)type {
+- (NSUInteger)type {
     return PKNodeTypeDelimited;
 }
