+++ branches/branch-dev-new-factory-2/test/nodes/PKPatternNode.m
@@ -43,7 +43,7 @@
 }
 
 
-- (int)type {
+- (NSUInteger)type {
     return PKNodeTypePattern;
 }
