+++ branches/branch-dev-new-factory-2/test/nodes/PKWhitespaceNode.m
@@ -10,7 +10,7 @@
 
 @implementation PKWhitespaceNode
 
-- (int)type {
+- (NSUInteger)type {
     return PKNodeTypeWhitespace;
 }
